+++ opentelemetry-api/CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## Unreleased
+### Metrics
+- Add instrument validation to `InstrumentBuilder`
+
 ## v0.18.0
 
 - API split from `opentelemetry` crate